+++ packages/gatsby/src/schema/infer/__tests__/infer-input.js
@@ -10,6 +10,22 @@ const { actions } = require(`../../../redux/actions`)
 const createPageDependency = require(`../../../redux/actions/add-page-dependency`)
 require(`../../../db/__tests__/fixtures/ensure-loki`)()
 
+jest.mock(`gatsby-cli/lib/reporter`, () => {
+  return {
+    log: jest.fn(),
+    info: jest.fn(),
+    warn: jest.fn(),
+    error: jest.fn(),
+    activityTimer: () => {
+      return {
+        start: jest.fn(),
+        setStatus: jest.fn(),
+        end: jest.fn(),
+      }
+    },
+  }
+})
+
 const buildTestSchema = async nodes => {
   store.dispatch({ type: `DELETE_CACHE` })
   nodes.forEach(node =>