+++ packages/gatsby/src/utils/webpack.config.js
@@ -23,7 +23,13 @@ const hasLocalEslint = require(`./local-eslint-config-finder`)
 //   3) build-javascript: Build JS and CSS chunks for production
 //   4) build-html: build all HTML files
 
-module.exports = async (program, directory, suppliedStage) => {
+module.exports = async (
+  program,
+  directory,
+  suppliedStage,
+  port,
+  { parentSpan } = {}
+) => {
   const modulesThatUseGatsby = await getGatsbyDependents()
   const directoryPath = withBasePath(directory)
 
@@ -560,6 +566,7 @@ module.exports = async (program, directory, suppliedStage) => {
     rules,
     loaders,
     plugins,
+    parentSpan,
   })
 
   return getConfig()